app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.26.0
yfinance>=0.2.0
redis>=5.0.0
requests>=2.31.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
apscheduler>=3.10.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2

# Email & Notifications
fastapi-mail>=1.4.1

# Data Export
openpyxl>=3.1.0
reportlab>=4.0.0

# WebSocket for real-time updates
websockets>=12.0
